+++ lib/dart/exceptions.dart
@@ -1,0 +1,11 @@
+class NotLoggedInException implements Exception {}
+
+class InvalidServerResponseException implements Exception {
+  String reason;
+  InvalidServerResponseException({required this.reason});
+
+  @override
+  String toString() {
+    return reason;
+  }
+}